Add remove examples.
